requirements.txt
django==5.0 
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.1.0

